Rename "activation" to "match" (and related terms) throughout
This renaming is to get closer to the terminology that seems to be
used in papers about production rules. "Activation" as a concept is
more about when a Condition or Rule is checking a condition, whereas
"matching" is when an object passes the requirements for a Condition
or Rule. This isn't a big deal to change at this early stage, so let's
just change it.