Avoid logically relying on database uniqueness constraints
Unfortunately, violating a uniqueness constraint causes Postgres to
fail the whole transaction in a way that we can't really recover from.
Unfortunately this means that we need to use some extra queries to
check before writing to the database (and introduce a race condition
in the process).

This isn't ideal, but it will have to do for now.