Significantly refactor Rule to remove "arguments" database field
This works by reconstructing the arguments from the relevant
ExtractorMatch records when we need to. This is a bit of a trade-off
of CPU against storage, but I think it's a better decision overall.

That said, this change is not as nice as it could be. Further
refactoring will be needed.