Add support for "on update" rule bodies
These allow for rules to be more efficiently updated in cases where we
know the "on match" clause has already fired. These bodies are passed
arguments in Struct objects containing "old" and "new" values. Thus,
the new value of a value bound with <name> can be accessed by using
name.new, and likewise the old value with name.old.