Add an execution_context global parameter for rule execution
The main purpose of this is to provide a way to define helper methods
that can be used in rule bodies. The mechanism in this commit is as
small an implementation as I could think of. There may be more
sophisticated things we could do that are helpful (e.g. provide the
Proc some context arguments), but it's not clear yet what direction is
best for this.