Make it possible to load rules from files
This change ended up being bigger than I expected. Doing the diff of
which rules are being defined/redefined means that we need to parse
the rules, which meant that we don't have the full rule string any
more. This required modifications to Rule to store the different
parts (and reconstruct some of them from the parsed rule!).

I'm relatively happy with the end result, but this diff is a nightmare.