Make identical rule definition not do any actual work
This is to pave the way for us to automatically load files from a
directory as part of our conventions. This will make it much easier to
integrate with a Rails application - if you can just add rules into a
file with a given extension (TBD) and have them get picked up
automatically that would be quite convenient.

The next step after that might be to hook into the db:migrate command,
or similar, to load the rules after each database migration.