Refactor parser to user richer AST objects
This is a pretty extensive change, but it makes it easier to add new
parts to the AST in future. There's still some refactoring necessary
beyond this to make it pleasant to work with, but this is a start.

The eventual goal of this change is to support grouping operators.