Fix support for top-level comparisons
This was added (incidentally) in the prior refactor, but it didn't
work properly. Having this makes it possible to express constraints
between different record matches at the top-level, rather than
requiring it to be done within a record matching clause.

This will also make it easier to use grouping operations in a way that
feels sensible. This example shows a top-level condition being used to
bind a value from a count grouping expression:

  User(<id>)
  <post_count> = count { Post(author_id = <id>) }

I haven't decided on the syntax that I want to use for this, but
that's the general idea.