Calculate bound values in Ruby during rule body execution
This isn't ideal, but it's the most straightforward way to make
top-level bindings work within the current framework. This creates a
semantic mismatch (because the logic is sometimes run in Ruby, and
sometimes run in SQL) and an efficiency problem (particularly if/when
we add grouping operations).

The next step for this will be to calculate these values as part of
the rule checking query, which will use more storage space for better
semantics and performance.