Trigger in after_{create,update,destroy} hooks only for relevant changes
This particular commit is a bit hacky, and it doesn't do all that I
want it to yet, but the general idea is there. This is fairly naive,
but fairly effective for "normal" queries.

Making this more efficient for negated and aggregated queries will be
more fun - we'll need to connect ids together through the variables to
identify which top-level objects to invalidate.