Factor out table edge calculation to be done one per rule definition
We can do some upfront work to make it so building the "changed ids"
query is faster. Essentially, we just construct the graph of fields
ahead of time in structures with faster lookup. Hopefully the searches
will be small enough that this doesn't make a huge difference, but
given how straightforward it is I've just done it.