Add support for "flags" on record fields, and use for enums
This change adds more logic around the way record fields are accessed,
and makes it possible for us to add "flags" to an attribute name in a
query. This allows us to ask for either form of an enum without
needing drastic changes. Now, you can use field:s for the string form,
and field:i for the integer form.