Allow rules to specify their timing, rather than the record class
This allows the same system to have some rules that fire in an
after_save hook, while others fire in after_commit hooks, and still
others fire in an ActiveJob. This makes the rules more flexible, and
makes it easier to handle all the setup in the library, which
simplifies project setup. Win-win!