Fetch match state immediately prior to execution
This change is needed to ensure that "stale" updates aren't run, and
then counted as "up to date" during an "after save" or "after commit"
rule. This has a performance cost, but it seems to be necessary for
now to avoid the behaviour being wrong.

I would like to improve this in future, but this cost should be less
significant for async rules, because they aren't quite so time
sensitive.