Track "edges" for each subquery separately when finding affected records
This allows us to put together a more precise query to select the
specific values that are affected, without different aggregates and/or
negations causing paths to be lost.

The code to do this search is well overdue for a refactor, but I'm not
doing that as part of this change.